refactor: add simulation module, split BaseSimulation and Simulation
