+++ src/pybamm/simulation/__init__.py
@@ -1,0 +1,2 @@
+from .base_simulation import BaseSimulation, load_sim, is_notebook
+from .simulation import Simulation
